+++ src/main/java/a2/Pizza/Pizza.java
@@ -7,7 +7,7 @@ import java.util.List;
  * Created by melo45 on 2019-03-21.
  */
 public interface Pizza {
-  void setPrice(int price);
+  void setPrice(double price);
 
   double getPrice();
